requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
pypdf
tqdm
openai
python-dotenv
beautifulsoup4
pyarrow
scikit-learn
click
rich
